+++ eviction_tracker/app.py
@@ -1,13 +1,14 @@
 import flask
 from flask import Flask, render_template, request, redirect
 from flask_security import hash_password, auth_token_required
-from eviction_tracker.extensions import assets, db, marshmallow, migrate, api, login_manager, user_datastore, security, User
+from eviction_tracker.extensions import assets, db, marshmallow, migrate, api, login_manager, security
+from eviction_tracker.admin.models import User, user_datastore
 import yaml
 import os
 import time
 
 from sqlalchemy import and_, or_, func, desc
-from eviction_tracker import commands, detainer_warrants
+from eviction_tracker import commands, detainer_warrants, admin, direct_action
 import json
 from datetime import datetime, date, timedelta
 from dateutil.rrule import rrule, MONTHLY
@@ -241,10 +242,10 @@ def register_extensions(app):
                      detainer_warrants.views.DetainerWarrantResource, app=app)
     api.add_resource('/phone-number-verifications/', detainer_warrants.views.PhoneNumberVerificationListResource,
                      detainer_warrants.views.PhoneNumberVerificationResource, app=app)
-    api.add_resource('/users/', detainer_warrants.views.UserListResource,
-                     detainer_warrants.views.UserResource, app=app)
-    api.add_resource('/roles/', detainer_warrants.views.RoleListResource,
-                     detainer_warrants.views.RoleResource, app=app)
+    api.add_resource('/users/', admin.views.UserListResource,
+                     admin.views.UserResource, app=app)
+    api.add_resource('/roles/', admin.views.RoleListResource,
+                     admin.views.RoleResource, app=app)
 
     @app.route('/api/v1/rollup/detainer-warrants')
     def detainer_warrant_rollup_by_month():
