+++ eviction_tracker/database.py
@@ -1,5 +1,5 @@
 """Database module, including the SQLAlchemy database object and DB-related utilities."""
-from sqlalchemy import text
+from sqlalchemy import text, func
 from sqlalchemy.orm import relationship
 from .extensions import db
 
@@ -18,3 +18,9 @@ def reference_col(tablename, nullable=False, pk_name='id', **kwargs):
     return db.Column(
         db.ForeignKey('{0}.{1}'.format(tablename, pk_name)),
         nullable=nullable, **kwargs)
+
+
+class Timestamped():
+    created_at = Column(db.DateTime, nullable=False, server_default=func.now())
+    updated_at = Column(
+        db.DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
