+++ eviction_tracker/detainer_warrants/views.py
@@ -22,65 +22,9 @@ from sqlalchemy.orm import raiseload
 
 from eviction_tracker.database import db
 from .models import DetainerWarrant, Attorney, Defendant, Courtroom, Plantiff, Judge, PhoneNumberVerification
-from eviction_tracker.extensions import User, Role, user_datastore
+from eviction_tracker.admin.models import User, Role, user_datastore
 from .serializers import *
-
-
-class CursorPagination(RelayCursorPagination):
-    def get_limit(self):
-        return 100
-
-    def get_page(self, query, view):
-        items = super().get_page(query, view)
-
-        after_cursor = None
-        if len(items) == self.get_limit():
-            after_cursor = super().make_cursor(
-                items[-1], view, super().get_field_orderings(view))
-
-        meta.update_response_meta({"after_cursor": after_cursor})
-        return items
-
-
-class Protected(AuthorizeModifyMixin, HasCredentialsAuthorizationBase):
-    @property
-    def request_user_id(self):
-        return self.get_request_credentials()["user_id"]
-
-    def filter_query(self, query, view):
-        return query
-
-    def authorize_modify_item(self, item, action):
-        user_id = self.request_user_id
-
-        if not user_id:
-            raise ApiError(403, {"code": "invalid_user"})
-
-
-class HeaderUserAuthentication(HeaderAuthenticationBase):
-    def get_credentials_from_token(self, token):
-        return {"user_id": token}
-
-
-class AllowDefendant(AuthorizeModifyMixin, HasCredentialsAuthorizationBase):
-    @property
-    def request_user_id(self):
-        return self.get_request_credentials()["user_id"]
-
-    def filter_query(self, query, view):
-        # viewer = user_datastore.find_user(id=self.request_user_id)
-
-        try:
-            return query.join(DetainerWarrant.defendants).filter(or_(
-                Defendant.name.ilike(f'%{current_user.first_name}%'),
-                Defendant.name.ilike(f'%{current_user.last_name}%')
-            ))
-        except AttributeError:
-            raise ApiError(403, {"code": "not a defendant"})
-
-    def authorize_modify_item(self, item, action):
-        if not self.request_user_id:
-            raise ApiError(403, {"code": "invalid_user"})
+from eviction_tracker.permissions.api import HeaderUserAuthentication, Protected, OnlyMe, CursorPagination, AllowDefendant
 
 
 class AttorneyResourceBase(GenericModelView):
@@ -237,60 +181,5 @@ class PhoneNumberVerificationListResource(PhoneNumberVerificationResourceBase):
 
 
 class PhoneNumberVerificationResource(PhoneNumberVerificationResourceBase):
-    def get(self, id):
-        return self.retrieve(id)
-
-
-class OnlyMe():
-    @property
-    def request_user_id(self):
-        return self.get_request_credentials()["user_id"]
-
-    def filter_query(self, query, view):
-        return query.filter_by(id=self.request_user_id)
-
-    def authorize_modify_item(self, item, action):
-        if not self.request_user_id:
-            raise ApiError(403, {"code": "invalid_user"})
-
-
-class UserResourceBase(GenericModelView):
-    model = User
-    schema = user_schema
-
-    authentication = HeaderUserAuthentication()
-    authorization = Protected()
-
-    pagination = CursorPagination()
-    sorting = Sorting('id', default='-id')
-
-
-class UserListResource(UserResourceBase):
-    def get(self):
-        return self.list()
-
-
-class UserResource(UserResourceBase):
-    def get(self, id):
-        return self.retrieve(id)
-
-
-class RoleResourceBase(GenericModelView):
-    model = Role
-    schema = role_schema
-
-    authentication = HeaderUserAuthentication()
-    authorization = Protected()
-
-    pagination = CursorPagination()
-    sorting = Sorting('id', default='-id')
-
-
-class RoleListResource(RoleResourceBase):
-    def get(self):
-        return self.list()
-
-
-class RoleResource(RoleResourceBase):
     def get(self, id):
         return self.retrieve(id)
