+++ eviction_tracker/direct_action/__init__.py
@@ -1,0 +1,2 @@
+"""The Direct Action module."""
+from . import views, models
