+++ eviction_tracker/extensions.py
@@ -38,54 +38,6 @@ class CRUDMixin(Model):
 
 
 db = SQLAlchemy(model_class=CRUDMixin)
-Column = db.Column
-
-
-roles_users = db.Table(
-    'roles_users', db.metadata,
-    db.Column('user_id',
-              db.ForeignKey('user.id'), primary_key=True),
-    db.Column('role_id', db.ForeignKey('role.id'), primary_key=True))
-
-
-class Role(db.Model, RoleMixin):
-    __tablename__ = 'role'
-    id = Column(db.Integer(), primary_key=True)
-    name = Column(db.String(80), unique=True)
-    description = Column(db.String(255))
-    users = relationship('User', secondary=roles_users, back_populates='roles')
-
-    def __repr__(self):
-        return f"<Role(name='{self.name}')>"
-
-
-class User(db.Model, UserMixin):
-    __tablename__ = 'user'
-    id = Column(db.Integer, primary_key=True)
-    email = Column(db.String(255), unique=True)
-    first_name = Column(db.String(255), nullable=False)
-    last_name = Column(db.String(255), nullable=False)
-    password = Column(db.String(255), nullable=False)
-    last_login_at = Column(db.DateTime())
-    current_login_at = Column(db.DateTime())
-    last_login_ip = Column(db.String(100))
-    current_login_ip = Column(db.String(100))
-    login_count = Column(db.Integer)
-    active = Column(db.Boolean())
-    fs_uniquifier = Column(db.String(255), unique=True, nullable=False)
-    confirmed_at = Column(db.DateTime())
-    roles = relationship('Role',
-                         secondary=roles_users,
-                         back_populates='users'
-                         )
-
-    @property
-    def name(self):
-        return self.first_name + ' ' + self.last_name
-
-    def __repr__(self):
-        return f"<User(name='{self.name}', email='{self.email}')>"
-
 
 # fsqla.FsModels.set_db_info(db)
 
@@ -98,5 +50,4 @@ login_manager = LoginManager()
 marshmallow = Marshmallow()
 migrate = Migrate(compare_type=True)
 api = Api(prefix='/api/v1')
-user_datastore = SQLAlchemyUserDatastore(db, User, Role)
 security = Security()
